+++ ClientApp/src/App.test.tsx
@@ -1,4 +1,3 @@
-import * as React from "react";
 import * as ReactDOM from "react-dom";
 import { MemoryRouter } from "react-router-dom";
 import App from "./App";
