+++ ClientApp/src/components/admin/completedConnections/ConnectionStyles.tsx
