+++ ClientApp/src/components/admin/completedConnections/Styles.tsx
