+++ ClientApp/src/components/admin/connectionSuggester/GiverTable.tsx
