+++ ClientApp/src/components/admin/dashboard/ConnectButton.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@material-ui/core";
 import React from "react";
 import useStyles from "./Styles";
-import { SelectedConnectionType } from "../../../common/components/Types";
+import { SelectedConnectionType } from "../../../components/shared/Types";
 
 type ConnectButtonProps = {
   selectedConnection: SelectedConnectionType;
