+++ ClientApp/src/components/admin/dashboard/Recipient.tsx
@@ -3,7 +3,7 @@ import { Search } from "@material-ui/icons";
 import React, { useState } from "react";
 import useStyles from "./Styles";
 import Datatable from "../common/RecipientTable";
-import { RecipientType} from "../../../common/components/Types";
+import { RecipientType} from "../../../components/shared/Types";
 
 type Props = {
   data: RecipientType[] | [];
