+++ ClientApp/src/components/admin/dashboard/Statistics.tsx
@@ -1,6 +1,6 @@
 import { Grid, Typography } from "@material-ui/core";
 import React, { useState, useEffect } from "react";
-import { GiverType, RecipientType } from "../../../common/components/Types";
+import { GiverType, RecipientType } from "../../../components/shared/Types";
 import { FiberManualRecord } from "@material-ui/icons";
 import useStyles from "./Styles";
