+++ ClientApp/src/components/shared/Types.tsx
@@ -1,4 +1,4 @@
-import Gender from "../enums/Gender";
+import Gender from "./../../common/enums/Gender"
 
 export interface SelectedConnectionType {
     giver?: GiverType;
