+++ ClientApp/src/pages/ErrorPage.tsx
@@ -1,4 +1,3 @@
-import * as React from "react";
 import { Container, Typography } from "@material-ui/core";
 
 interface Props {
